Fix docker url
Ubuntu did a weird thing where they decided to create a dns address for `docker.for.mac.localhost` which breaks small-prox on people using ubuntu. Luckily, they didn't also use the alias which is `docker.for.mac.host.internal`, so this change uses that one instead, which should act the same.